Align marketing controller tests with controller structure
