+++ tests/Feature/Controllers/Marketing/Docs/Api/AccountControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Controllers\Marketing;
+namespace Tests\Feature\Controllers\Marketing\Docs\Api;
 
 use App\Jobs\RecordMarketingPageVisit;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -10,22 +10,22 @@ use Illuminate\Support\Facades\Queue;
 use PHPUnit\Framework\Attributes\Test;
 use Tests\TestCase;
 
-final class ProfileControllerTest extends TestCase
+final class AccountControllerTest extends TestCase
 {
     use RefreshDatabase;
 
     #[Test]
-    public function it_renders_the_profile_api_docs_page(): void
+    public function it_renders_the_account_api_docs_page(): void
     {
         Queue::fake();
 
-        $response = $this->get(route('marketing.docs.api.account.profile', absolute: false));
+        $response = $this->get(route('marketing.docs.api.account', absolute: false));
 
         $response->assertOk();
-        $response->assertViewIs('marketing.docs.api.account.profile');
+        $response->assertViewIs('marketing.docs.api.account.account');
 
         Queue::assertPushedOn('low', RecordMarketingPageVisit::class, function (RecordMarketingPageVisit $job): bool {
-            return $job->viewName === 'marketing.docs.api.account.profile';
+            return $job->viewName === 'marketing.docs.api.account.prune';
         });
     }
 }
