+++ tests/Feature/Controllers/Marketing/Docs/Api/ApiIntroductionControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Controllers\Marketing;
+namespace Tests\Feature\Controllers\Marketing\Docs\Api;
 
 use App\Jobs\RecordMarketingPageVisit;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -10,22 +10,22 @@ use Illuminate\Support\Facades\Queue;
 use PHPUnit\Framework\Attributes\Test;
 use Tests\TestCase;
 
-final class ApiManagementControllerTest extends TestCase
+final class ApiIntroductionControllerTest extends TestCase
 {
     use RefreshDatabase;
 
     #[Test]
-    public function it_renders_the_api_management_page(): void
+    public function it_renders_the_api_introduction_page(): void
     {
         Queue::fake();
 
-        $response = $this->get(route('marketing.docs.api.account.api-management', absolute: false));
+        $response = $this->get(route('marketing.docs.api.index', absolute: false));
 
         $response->assertOk();
-        $response->assertViewIs('marketing.docs.api.account.api-management');
+        $response->assertViewIs('marketing.docs.api.introduction');
 
         Queue::assertPushedOn('low', RecordMarketingPageVisit::class, function (RecordMarketingPageVisit $job): bool {
-            return $job->viewName === 'marketing.docs.api.account.api-management';
+            return $job->viewName === 'marketing.docs.api.introduction';
         });
     }
 }
