+++ tests/Feature/Controllers/Marketing/Docs/Api/AuthenticationControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Controllers\Marketing;
+namespace Tests\Feature\Controllers\Marketing\Docs\Api;
 
 use App\Jobs\RecordMarketingPageVisit;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -10,22 +10,22 @@ use Illuminate\Support\Facades\Queue;
 use PHPUnit\Framework\Attributes\Test;
 use Tests\TestCase;
 
-final class ApiIntroductionControllerTest extends TestCase
+final class AuthenticationControllerTest extends TestCase
 {
     use RefreshDatabase;
 
     #[Test]
-    public function it_renders_the_api_introduction_page(): void
+    public function it_renders_the_authentication_api_docs_page(): void
     {
         Queue::fake();
 
-        $response = $this->get(route('marketing.docs.api.index', absolute: false));
+        $response = $this->get(route('marketing.docs.api.authentication', absolute: false));
 
         $response->assertOk();
-        $response->assertViewIs('marketing.docs.api.introduction');
+        $response->assertViewIs('marketing.docs.api.authentication');
 
         Queue::assertPushedOn('low', RecordMarketingPageVisit::class, function (RecordMarketingPageVisit $job): bool {
-            return $job->viewName === 'marketing.docs.api.introduction';
+            return $job->viewName === 'marketing.docs.api.authentication';
         });
     }
 }
