+++ tests/Feature/Controllers/Marketing/Docs/Api/ProfileControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Controllers\Marketing;
+namespace Tests\Feature\Controllers\Marketing\Docs\Api;
 
 use App\Jobs\RecordMarketingPageVisit;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -10,22 +10,22 @@ use Illuminate\Support\Facades\Queue;
 use PHPUnit\Framework\Attributes\Test;
 use Tests\TestCase;
 
-final class DocControllerTest extends TestCase
+final class ProfileControllerTest extends TestCase
 {
     use RefreshDatabase;
 
     #[Test]
-    public function it_renders_the_docs_index(): void
+    public function it_renders_the_profile_api_docs_page(): void
     {
         Queue::fake();
 
-        $response = $this->get(route('marketing.docs.index', absolute: false));
+        $response = $this->get(route('marketing.docs.api.account.profile', absolute: false));
 
         $response->assertOk();
-        $response->assertViewIs('marketing.docs.api.introduction');
+        $response->assertViewIs('marketing.docs.api.account.profile');
 
         Queue::assertPushedOn('low', RecordMarketingPageVisit::class, function (RecordMarketingPageVisit $job): bool {
-            return $job->viewName === 'marketing.docs.api.introduction';
+            return $job->viewName === 'marketing.docs.api.account.profile';
         });
     }
 }
