+++ tests/Feature/Controllers/Marketing/Docs/Concepts/PermissionControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Controllers\Marketing;
+namespace Tests\Feature\Controllers\Marketing\Docs\Concepts;
 
 use App\Jobs\RecordMarketingPageVisit;
 use Illuminate\Foundation\Testing\RefreshDatabase;
